requirements.txt
annotated-types==0.6.0
autopep8==2.0.4
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
exceptiongroup==1.2.0
greenlet==3.0.3
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
instagrapi==2.0.1
moviepy==1.0.3
numpy==1.24.4
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
proglog==0.1.10
pycodestyle==2.11.1
pycryptodomex==3.18.0
pydantic==2.5.2
pydantic-core==2.14.5
PySocks==1.7.1
pytest==8.0.2
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.27
tomli==2.0.1
tqdm==4.66.2
typing-extensions==4.10.0
urllib3==2.2.1
